fix: address PR review — type safety, architecture, and SSE fixes
- Remove `result` state from useAgentStream (backend `done` event has no `result` field)
- Add encodeURIComponent to SSE URL construction for path safety
- Extract shared types (AgentStep, StepPhase, StreamEvent) to app/types/stream.ts
  to resolve hook→component reverse dependency
- Remove dead `sourceRef` code; consolidate to `cleanupRef`
- Add `setStatus("error")` on server error event reception
- Move api.ts from app/lib/ to lib/ to align with shadcn aliases
- Change `accent` prop to AccentColor union type with internal class mapping
- Narrow `icon` prop from ReactNode to ReactElement
- Fix --font-sans CSS self-reference with --font-geist-sans fallback
- Add NEXT_PUBLIC_API_URL to .env.example
