+++ .env.example
@@ -12,3 +12,6 @@ QDRANT_PORT=6333
 
 # Rate Limiting
 DAILY_LIMIT=50
+
+# Frontend
+NEXT_PUBLIC_API_URL=http://localhost:8000
