+++ frontend/app/globals.css
@@ -74,7 +74,7 @@
 }
 
 @theme inline {
-    --font-sans: var(--font-sans);
+    --font-sans: var(--font-geist-sans, ui-sans-serif, system-ui, sans-serif);
     --color-sidebar-ring: var(--sidebar-ring);
     --color-sidebar-border: var(--sidebar-border);
     --color-sidebar-accent-foreground: var(--sidebar-accent-foreground);
